backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
dependency-injector>=4.41.0
python-dotenv>=1.0.0

# AI Models
anthropic>=0.39.0
zhipuai>=2.0.0
google-genai>=1.0.0
tiktoken>=0.5.0
